requirements.txt
Flask
Flask-SocketIO
argon2-cffi
python-dotenv
qrcode
rich
gevent
pyperclip
google-generativeai
